+++ index.js
@@ -1,299 +1,349 @@
 import { readFileSync } from 'fs';
 import { createServer } from 'http';
 import { createSecureServer } from 'http2';
+import { dirname, resolve } from 'path';
+import { fileURLToPath } from 'url';
 
 import Koa from 'koa';
 
-import Compress from 'koa-compress';
-import { constants } from 'zlib';
-import Cors from '@koa/cors';
-import Helmet from 'koa-helmet';
-import Favicon from 'koa-favicon';
 
-import context from 'koa/lib/context';
-import response from 'koa/lib/response';
-import request from 'koa/lib/request';
+import KoaCompress from 'koa-compress';
+import KoaCors from '@koa/cors';
+import KoaHelmet from 'koa-helmet';
+import KoaFavicon from 'koa-favicon';
 
-import { T, TT } from './lib/i18n.js';
+import KoaContext from 'koa/lib/context';
+import KoaResponse from 'koa/lib/response';
+import KoaRequest from 'koa/lib/request';
 
+import { loadI18NResource, TT } from '@nuogz/i18n';
+import { injectBaseLogger } from '@nuogz/utility';
+
+
+
+/** @typedef {import('@nuogz/utility/src/injectBaseLogger.js').LoggerLike} LoggerLike */
+/** @typedef {import('@nuogz/utility/src/injectBaseLogger.js').LoggerOption} LoggerOption */
 
 
 /**
- * - simple server library based on koa
- * @class
+ * should equivalent to the first argument of `new Koa(option)`
+ * @typedef {Object} KoaOption
+ * @property {string} [env]
+ * @property {string[]} [keys]
+ * @property {boolean} [proxy]
+ * @property {number} [subdomainOffset]
+ * @property {string} [proxyIpHeader]
+ * @property {number} [maxIpsCount]
  */
+
+/**
+ * should equivalent to the argument of `new KoaFavicon(path, { maxage })`
+ * @typedef {Object} KoaFaviconOption
+ * @property {string} path
+ * @property {number} [maxage] ms
+ */
+
+
+/**
+ * @typedef {Object} DesireExtendDisableOption
+ * @property {boolean} [disable]
+ */
+
+/**
+ * @typedef {Object} ModuleOption
+ *
+ * @property {import('http').ServerOptions} [http] node module HTTP option
+ * @property {import('http2').SecureServerOptions & DesireExtendDisableOption} [http2] node module HTTP2 option
+ *
+ * @property {KoaOption} [koa] module `koa` option
+ *
+ * @property {KoaCompress.CompressOptions & DesireExtendDisableOption} [compress] module `koa-compress` option
+ * @property {KoaCors.Options & DesireExtendDisableOption} [cors] module `@koa/cors` option
+ * @property {KoaHelmet.KoaHelmetContentSecurityPolicyConfiguration & DesireExtendDisableOption} [csp] module `koa-helmet` contentSecurityPolicy option
+ * @property {(KoaFaviconOption & DesireExtendDisableOption) | string} [favicon] module `koa-favicon` option or favicon path
+ */
+
+
+/**
+ * Desire constructor option
+ * @typedef {Object} DesireOption
+ *
+ * @property {string} [name] name for server. Used to log output
+ *
+ * @property {string} [host] listen host
+ * @property {number} [port] listen port
+ *
+ * @property {ModuleOption} [module={}]
+ *
+ * @property {Object | import('@nuogz/desire-harbour').HarbourOption} [harbour] Harbour Option
+ * @property {Function | string} [Harbour] the interface and folder mapping initializer, called `Harbour`, which is used to apply options to the `koajs` instance, and is invoked by passing an instance of `koajs`. pass string `'default'`, `''` or undefined will use module `@nuogz/desire-harbour`; pass a `string` will try to import a module with the same name as the option；pass a `class` will be created and then call its `init()` method; pass a `function` will be called directly
+ *
+ * @property {LoggerOption} [logger]
+ */
+
+
+
+loadI18NResource('@nuogz/desire', resolve(dirname(fileURLToPath(import.meta.url)), 'locale'));
+
+const T = TT('@nuogz/desire');
+
+
+const hasOption = (key, object) => key in object && object[key] !== undefined;
+
+
+
 export default class Desire {
-	static KoaContext = context;
-	static KoaResponse = response;
-	static KoaRequest = request;
+	/** @type {Koa.Context} */
+	static KoaContext = KoaContext;
+	/** @type {Koa.Response} */
+	static KoaResponse = KoaResponse;
+	/** @type {Koa.Request} */
+	static KoaRequest = KoaRequest;
 
 
-	/**
-	 * Face config
-	 * @typedef {Object} FaceConfig
-	 * @property {string} method Multi methods splited by `;`. methods used in `koa-router` (HTTP 1.1), or `wock` for wock face
-	 * @property {string} route
-	 * @property {Function} handle
-	 * @property {boolean} upload Indicates whether file upload is enabled. `undefined` or `true` for enabled；`false` for disabled
-	 */
 
-	/**
-	 * Folder mapping config
-	 * @typedef {Object} FolderConfig
-	 * @property {string} route **ATTENTION** This option is fully independent that will not concat with `{ServerConfig.facePrefix}`
-	 * @property {string} path path of file system
-	 * @property {Object} option `koa-mount` config
-	 */
-
-	/**
-	 * Logger config
-	 * - include `trace`, `debug`, `info`, `warn`, `error`, `fatal` and `mark` for classify log output
-	 * @typedef {Object} LoggerConfig
-	 * @property {Function} trace
-	 * @property {Function} debug
-	 * @property {Function} info
-	 * @property {Function} warn
-	 * @property {Function} error
-	 * @property {Function} fatal
-	 * @property {Function} mark
-	 */
-
-	/**
-	 * Mare(Middleware) Initial Function
-	 * - Promise is supported.
-	 * @typedef {Function} MareIniter
-	 * @async
-	 */
-	/**
-	 * Mare(Middleware) config
-	 * - Mare, abbreviation for `Middleware`.
-	 * - Its route is the same as Face by default.
-	 * @typedef {Object} MareConfig
-	 * @property {Array<MareIniter>|string} [before] An array for before-mare initer, or a string for one built-in mare.
-	 * @property {Array<MareIniter>|string} [after] An array for after-mare initer, or a string for one built-in mare.
-	 */
-	/**
-	 * Mare(Middleware) config for Wock.
-	 * - Its route is the same as Face by default.
-	 * @typedef {Object} WockMareConfig
-	 * @property {Array<MareIniter>|string} [before] An array for before-mare initer, or a string for one built-in mare.
-	 * @property {Array<MareIniter>|string} [after] An array for after-mare initer, or a string for one built-in mare.
-	 * @property {Array<MareIniter>|string} [upgrade] An array for upgrade-mare initer, or a string for one built-in mare.
-	 * @property {Array<MareIniter>|string} [close] An array for close-mare initer, or a string for one built-in mare.
-	 */
-
-	/**
-	 * Wock config
-	 * - Wock, abbreviation for `WebSocket`.
-	 * - Its route is the same as Face by default.
-	 * @typedef {Object} WockConfig
-	 * @property {boolean} [disable=false] Indicates whether disabled. `undefined` or `true` for disabled；`false` for enabled.
-	 * @property {string} [route='wock'] Route under WebSocket. **ATTENTION** This option is fully independent that will not concat with `{ServerConfig.facePrefix}`
-	 * @property {boolean} [ping=false] Indicates whether send `ping` event after websocket connected. `undefined` or `false` for not send；`true` for will send
-	 * @property {WockMareConfig} [mare] Mare(Middleware) config for Wock.
-	 */
-
-	/**
-	 * HTTP2 config
-	 * @typedef {Object} HTTP2Config
-	 * @property {boolean} [disable=false] Indicates whether disabled. `undefined` or `true` for disabled；`false` for enabled.
-	 * @property {boolean} [http1=true] Indicates whether allow HTTP1 request
-	 * @property {string|Buffer} key SSL private key. `string` for file path
-	 * @property {string|Buffer} cert SSL public cert. `string` for file path
-	 */
-
-	/**
-	 * Server config
-	 * @typedef {Object} ServerConfig
-	 * @property {string} [name] name for server. Used to log output
-	 *
-	 * @property {string} host listen host
-	 * @property {number} post listen port
-	 *
-	 * @property {string} [facePrefix='/'] prefix for Faces
-	 * @property {Array<FaceConfig>} [faces] An array for Face Config
-	 * @property {Array<FolderConfig>} [folds] An array for Folder Mapping
-	 * @property {MareConfig} [mare] Mare config
-	 * @property {WockConfig} [wock] Wock config
-	 * @property {Function|string} [harb] Harb(harbour) initial function. `undefined` or `'default'` for using default Harb library `@nuogz/desire-harb-default`; `string` for import library；`Function` for Harb self. Harb is used to apply configs into `koajs` instance. only `Desire` instances will be passed.
-	 *
-	 * @property {HTTP2Config|boolean} [http2=false] HTTP2 config
-	 *
-	 * @property {Function|LoggerConfig} [logger] Logger config. `undefined` for use `console` functions；`false` for close output；`Function` for output non-classify logs；`{LoggerConfig}` for classify logs. The function will be called in the format of where, what and result. **ATTENTION** The Error instance will be passed in as one of the result arguments, not stringified error text.
-	 *
-	 * @property {string} [favicon] path of icon file
-	 *
-	 * @property {string} [locale] locale for log
-	 *
-	 * @property {false|Compress.CompressOptions} [compress] compress options
-	 *
-	 */
-
-	/**
-	 * @param {ServerConfig} C server main config
-	 */
-	constructor(C) {
-		if(!C) { throw TypeError(T('error.missServerConfig')); }
-
-		/**
-		 * server config
-		 * @type {ServerConfig}
-		 */
-		this.C = C;
-
-		this.locale = C.locale;
-		this.TT = TT(this.locale);
+	/** @type {DesireOption} */
+	optionRaw;
 
 
-		this.initLogger();
-		this.initHTTP2();
+	/** @type {import('http').Server | import('http2').Http2Server} */
+	server;
+	/** @type {Koa} */
+	koa;
 
-		this.server = this.http2 ? createSecureServer(this.http2) : createServer();
-		this.koa = new Koa();
 
-		this.initHeader();
-		this.initFavIcon();
+	/** @type {string} */
+	name = T('Server');
 
+
+	/** @type {string} */
+	host;
+	/** @type {number} */
+	port;
+
+
+
+	/** @type {ModuleOption} */
+	option = {};
+
+
+	/** @type {Object} */
+	optionHarbour = {};
+	/** @type {Object} */
+	harbour;
+	/** @type {Function|string} */
+	Harbour;
+	/** @type {Function|string} */
+	HarbourImport;
+
+
+
+	/** @type {LoggerLike} */
+	logTrace;
+	/** @type {LoggerLike} */
+	logDebug;
+	/** @type {LoggerLike} */
+	logInfo;
+	/** @type {LoggerLike} */
+	logError;
+	/** @type {LoggerLike} */
+	logWarn;
+	/** @type {LoggerLike} */
+	logFatal;
+	/** @type {LoggerLike} */
+	logMark;
+
+
+
+	/** @param {DesireOption} [option] */
+	constructor(option = {}) {
+		this.optionRaw = option;
+
+
+		this.name = hasOption('name', option) ? option.name : this.name;
+
+		this.host = hasOption('host', option) ? option.host : this.host;
+		this.port = hasOption('port', option) ? option.port : this.port;
+
+
+		this.option = hasOption('module', option) ? option.module : this.option;
+
+
+		this.optionHarbour = hasOption('harbour', option) ? option.harbour : this.optionHarbour;
+		this.Harbour = hasOption('Harbour', option) ? option.Harbour : this.Harbour;
+
+
+		injectBaseLogger(this, Object.assign({ name: this.name }, option.logger));
+
+
+
+		this.initBase();
+
+		this.initFavicon();
+
+		return this.initHeader()
+			.then(() => this);
 	}
 
-	/** server protocol */
-	get protocol() { return this.http2 ? 'http2' : 'http'; }
-	/** server name for log */
-	get nameLog() { return typeof this.C.name == 'string' ? this.C.name : 'Desire'; }
 
-	/** start server */
-	async start() {
-		const { C: { host, port }, server, logInfo, logFatal } = this;
+	initBase() {
+		const { option: { koa, http, http2 } } = this;
 
-		try {
-			await this.initHarbour();
-			await this.initServer();
+		const useHTTP2 = http2 && http2.disable !== false;
 
-			// listen port
-			await new Promise((resolve, reject) => server.listen(port, host, error => error ? reject(error) : resolve()));
 
-			// try to downgrade system perm after listen
-			try {
-				const env = process.env,
-					uid = parseInt(env['SUDO_UID'] || process.getuid(), 10),
-					gid = parseInt(env['SUDO_GID'] || process.getgid(), 10);
+		if(useHTTP2) {
+			http2.key = typeof http2.key == 'string'
+				? readFileSync(http2.key)
+				: http2.key;
+			http2.cert = typeof http2.cert == 'string'
+				? readFileSync(http2.cert)
+				: http2.cert;
+		}
 
-				process.setgid(gid);
-				process.setuid(uid);
+
+		this.server = useHTTP2 ? createSecureServer(http2) : createServer(http);
+
+
+		this.koa = new Koa(koa);
+	}
+
+	initFavicon() {
+		const { option: { favicon }, koa, logDebug } = this;
+
+
+		if(favicon && favicon?.disable !== false) {
+			if(typeof favicon == 'string') {
+				koa.use(KoaFavicon(favicon));
 			}
-			catch(e) { void 0; }
+			else {
+				koa.use(KoaFavicon(favicon.path, { maxage: favicon.maxage }));
+			}
 
-			logInfo(this.TT('listenPort', { url: `${this.protocol}://${host == '0.0.0.0' ? 'localhost' : host}:${port}` }), `✔ `);
+			logDebug(T('initFavicon'), T('initFaviconArgument', { favicon: favicon.path || favicon }));
 		}
-		catch(error) {
-			logFatal(this.TT('listenPort', { url: `${this.protocol}://${host}:${port}` }), error);
-		}
-
-		return this;
 	}
 
-	/** init HTTP2 */
-	initHTTP2() {
-		const { C: { http2 } } = this;
-
-		if(!http2 || http2.disable) { return false; }
-
-		const config = {
-			allowHTTP1: http2.http1,
-			key: http2.key,
-			pem: http2.pem,
-		};
-
-		if(typeof config.key == 'string') { config.key = readFileSync(config.key); }
-		if(typeof config.pem == 'string') { config.pem = readFileSync(config.pem); }
-
-		this.http2 = config;
-	}
+	async initHeader() {
+		const { option: { compress, cors, csp }, koa } = this;
 
 
-	/** init HTTP header */
-	initHeader() {
-		const koa = this.koa;
+		// ZLIB compress
+		if(compress && compress.disable !== false) {
+			const { constants } = await import('zlib');
 
-		// zlib
-		if(this.C.compress !== false) {
-			koa.use(Compress(Object.assign({}, this.C.compress, {
+			koa.use(KoaCompress(Object.assign({}, compress, {
 				threshold: 2048,
 				gzip: { flush: constants.Z_SYNC_FLUSH },
 				deflate: { flush: constants.Z_SYNC_FLUSH },
 			})));
 		}
 
-		// cors header
-		koa.use(Cors());
 
-		// hsts header
-		koa.use(Helmet.contentSecurityPolicy({
-			directives: {
-				defaultSrc: ['\'self\'', '\'unsafe-inline\'', '\'unsafe-eval\''],
-				objectSrc: ['\'none\''],
-				scriptSrc: ['\'self\'', '\'unsafe-inline\'', '\'unsafe-eval\''],
-				styleSrc: ['\'self\'', 'https:', '\'unsafe-inline\''],
-				imgSrc: ['\'self\'', 'https:', 'data:', 'blob:', 'mediastream:', 'filesystem:'],
-				fontSrc: ['\'self\'', 'https:', 'data:', 'blob:', 'mediastream:', 'filesystem:'],
-			},
-		}));
-		koa.use(Helmet.expectCt());
-		koa.use(Helmet.frameguard());
-		koa.use(Helmet.hidePoweredBy());
-		koa.use(Helmet.hsts());
-		koa.use(Helmet.ieNoOpen());
-		koa.use(Helmet.noSniff());
-		koa.use(Helmet.permittedCrossDomainPolicies());
-		koa.use(Helmet.referrerPolicy());
-		koa.use(Helmet.xssFilter());
-	}
+		// CORS header
+		if(cors?.disable !== false) {
+			koa.use(KoaCors(cors));
+		}
 
-	/** init Favicon */
-	initFavIcon() {
-		const { C: { favicon }, koa, logDebug } = this;
 
-		if(favicon && typeof favicon == 'string') {
-			koa.use(Favicon(favicon));
+		// HSTS header
+		if(csp?.disable !== false) {
+			koa.use(KoaHelmet.contentSecurityPolicy(Object.assign({}, csp, {
+				directives: {
+					defaultSrc: ['\'self\'', '\'unsafe-inline\'', '\'unsafe-eval\''],
+					objectSrc: ['\'none\''],
+					scriptSrc: ['\'self\'', '\'unsafe-inline\'', '\'unsafe-eval\''],
+					styleSrc: ['\'self\'', 'https:', '\'unsafe-inline\''],
+					imgSrc: ['\'self\'', 'https:', 'data:', 'blob:', 'mediastream:', 'filesystem:'],
+					fontSrc: ['\'self\'', 'https:', 'data:', 'blob:', 'mediastream:', 'filesystem:'],
+				},
+			})));
 
-			logDebug(this.TT('initFavicon'), this.TT('initFaviconParam', { favicon }));
+			koa.use(KoaHelmet.frameguard());
+			koa.use(KoaHelmet.hidePoweredBy());
+			koa.use(KoaHelmet.hsts());
+			koa.use(KoaHelmet.ieNoOpen());
+			koa.use(KoaHelmet.noSniff());
+			koa.use(KoaHelmet.permittedCrossDomainPolicies());
+			koa.use(KoaHelmet.referrerPolicy());
+			koa.use(KoaHelmet.xssFilter());
 		}
 	}
 
-	/**
-	 * init Harb
-	 * @async
-	 */
+
+
+	/** server protocol */
+	get protocol() { return typeof this.option.http2 == 'object' && this.option.http2 != null ? 'http2' : 'http'; }
+
+
+	/** start server */
+	async start() {
+		const { host, port, server, logFatal, logInfo } = this;
+
+
+		try {
+			await this.initHarbour();
+
+			this.initServer();
+
+			// listen port
+			await new Promise((resolve, reject) =>
+				server.listen(port, host, error => error ? reject(error) : resolve())
+			);
+
+
+			logInfo(T('listen', { url: `${this.protocol}://${host == '0.0.0.0' ? 'localhost' : host}:${port}` }), `✔ `);
+		}
+		catch(error) {
+			logFatal(T('listen', { url: `${this.protocol}://${host}:${port}` }), error);
+		}
+
+
+		return this;
+	}
+
+
+	/** init Harbour */
 	async initHarbour() {
-		const { C: { harb }, logInfo, logFatal } = this;
+		const { optionHarbour: option, logFatal, logInfo } = this;
+		let { Harbour } = this;
 
-		if(harb !== false) {
-			try {
-				if(typeof harb == 'function') {
-					this.harb = await harb(this);
-				}
-				else if(harb != 'default' && typeof harb == 'string') {
-					this.harb = await (await import(harb)).default(this);
-				}
-				else {
-					this.harb = await (await import('@nuogz/desire-harb-default')).default(this);
-				}
 
-				logInfo(this.TT('initHarb'), '✔ ');
+		if(option === false) { return; }
+
+
+		try {
+			if(Harbour != 'default' && Harbour != '' && typeof Harbour == 'string') {
+				Harbour = this.HarbourImport = (await import(Harbour)).default;
 			}
-			catch(error) {
-				logFatal(this.TT('initHarb'), error);
+			else {
+				Harbour = this.HarbourImport = (await import('@nuogz/desire-harbour')).default;
 			}
+
+
+			if(typeof Harbour == 'function' && Reflect.getOwnPropertyDescriptor(Harbour, 'prototype')) {
+				this.harbour = await new Harbour(this, option);
+
+				await this.harbour.init(this, option);
+			}
+			else if(typeof Harbour == 'function') {
+				this.harbour = await Harbour(this, option);
+			}
+			else {
+				throw Error(T('invalidHarbour', { value: Harbour }));
+			}
+
+			logInfo(T('initHarbour'), '✔ ');
+		}
+		catch(error) {
+			logFatal(T('initHarbour'), error);
 		}
 	}
 
 
-	/**
-	 * init server
-	 * @async
-	 */
-	async initServer() {
-		const { C: { host, port }, server, koa, logFatal } = this;
+	initServer() {
+		const { host, port, server, koa, logFatal } = this;
+
 
 		// request listen
 		server.on('request', koa.callback());
@@ -301,100 +351,13 @@ export default class Desire {
 		// listen error
 		server.on('error', error => {
 			if(error.code == 'EADDRINUSE') {
-				logFatal(this.TT('listenPort', { url: `${this.protocol}://${host}:${port}` }), this.TT('listenPortError'));
+				logFatal(T('listen', { url: `${this.protocol}://${host}:${port}` }), T('listenError'));
 			}
 			else {
-				logFatal(this.TT('listenPort', { url: `${this.protocol}://${host}:${port}` }), error);
+				logFatal(T('listen', { url: `${this.protocol}://${host}:${port}` }), error);
 			}
 
-			process.exit();
+			server.close();
 		});
-	}
-
-	/** init logger */
-	initLogger() {
-		const { C: { logger } } = this;
-
-		// log close
-		if(logger === false) {
-			this.logTrace = () => { };
-			this.logDebug = () => { };
-			this.logInfo = () => { };
-			this.logError = () => { };
-			this.logWarn = () => { };
-			this.logFatal = () => { };
-			this.logMark = () => { };
-		}
-		// log by single function
-		else if(typeof logger == 'function') {
-			this.logTrace = (...params) => logger(this.nameLog, ...params);
-			this.logDebug = (...params) => logger(this.nameLog, ...params);
-			this.logInfo = (...params) => logger(this.nameLog, ...params);
-			this.logError = (...params) => logger(this.nameLog, ...params);
-			this.logWarn = (...params) => logger(this.nameLog, ...params);
-			this.logFatal = (...params) => logger(this.nameLog, ...params);
-			this.logMark = (...params) => logger(this.nameLog, ...params);
-		}
-		// classify log
-		else {
-			const csl = console;
-
-			this.logTrace =
-				typeof logger?.trace == 'function' ?
-					(...params) => logger.trace(this.nameLog, ...params) :
-					(
-						typeof csl.trace == 'function' ?
-							(...params) => csl.trace(this.nameLog, ...params) :
-							(...params) => csl.log(this.nameLog, ...params)
-					);
-			this.logDebug =
-				typeof logger?.debug == 'function' ?
-					(...params) => logger.debug(this.nameLog, ...params) :
-					(
-						typeof csl.debug == 'function' ?
-							(...params) => csl.debug(this.nameLog, ...params) :
-							(...params) => csl.log(this.nameLog, ...params)
-					);
-			this.logInfo =
-				typeof logger?.info == 'function' ?
-					(...params) => logger.info(this.nameLog, ...params) :
-					(
-						typeof csl.info == 'function' ?
-							(...params) => csl.info(this.nameLog, ...params) :
-							(...params) => csl.log(this.nameLog, ...params)
-					);
-			this.logError =
-				typeof logger?.error == 'function' ?
-					(...params) => logger.error(this.nameLog, ...params) :
-					(
-						typeof csl.error == 'function' ?
-							(...params) => csl.error(this.nameLog, ...params) :
-							(...params) => csl.log(this.nameLog, ...params)
-					);
-			this.logWarn =
-				typeof logger?.warn == 'function' ?
-					(...params) => logger.warn(this.nameLog, ...params) :
-					(
-						typeof csl.warn == 'function' ?
-							(...params) => csl.warn(this.nameLog, ...params) :
-							(...params) => csl.log(this.nameLog, ...params)
-					);
-			this.logFatal =
-				typeof logger?.fatal == 'function' ?
-					(...params) => logger.fatal(this.nameLog, ...params) :
-					(
-						typeof csl.error == 'function' ?
-							(...params) => csl.error(this.nameLog, ...params) :
-							(...params) => csl.log(this.nameLog, ...params)
-					);
-			this.logMark =
-				typeof logger?.mark == 'function' ?
-					(...params) => logger.mark(this.nameLog, ...params) :
-					(
-						typeof csl.info == 'function' ?
-							(...params) => csl.info(this.nameLog, ...params) :
-							(...params) => csl.log(this.nameLog, ...params)
-					);
-		}
 	}
 }